Split Lapack trait definition into sub file
